build(Rollup): Replace basic TypeScript compiler with Rollup
Add Rollup
Rename tsconfig to include a default tsconfig.json
Add utils for named exports
Refactor JS util files to TS
Make VSCode use tsdk from project
