+++ .vscode/settings.json
@@ -6,5 +6,6 @@
     "eslint.validate": [
         "javascript",
         "typescript"
-    ]
+    ],
+    "typescript.tsdk": "node_modules/typescript/lib"
 }
