+++ package.json
@@ -19,16 +19,16 @@
   },
   "homepage": "https://github.com/isaaceindhoven/testcafe-reporter-allure#readme",
   "scripts": {
-    "build:prod": "tsc --project tsconfig.prod.json",
-    "build:test": "tsc --project tsconfig.test.json",
-    "test": "npm run test:lint && npm run test:unit && npm run build:prod && npm run test:e2e",
+    "prebuild": "rimraf dist",
+    "build": "rollup -c",
+    "test": "npm run test:lint && npm run test:unit && npm run test:e2e",
     "test:lint": "eslint --ext .js,.ts --ignore-pattern !.* --ignore-path .gitignore .",
     "test:lint:fix": "eslint --ext .js,.ts --ignore-pattern !.* --ignore-path .gitignore . --fix",
     "test:unit": "jest",
     "test:unit:watch": "jest --config=jest.config.js --watch",
+    "pretest:e2e": "npm run build",
     "test:e2e": "node tests/runner.js",
-    "test:e2e:build": "npm run build:prod && node tests/runner.js",
-    "test:e2e:build:allure": "npm run build:prod && node tests/runner.js && npm run allure",
+    "test:e2e:build:allure": "npm run test:e2e && npm run allure",
     "allure": "allure generate ./allure/allure-results --clean -o ./allure/allure-report && allure open ./allure/allure-report",
     "version": "node -v",
     "browsers": "testcafe --list-browsers",
@@ -49,7 +49,9 @@
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
+    "@rollup/plugin-typescript": "^4.1.2",
     "@types/jest": "^25.2.1",
+    "@types/rimraf": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^2.28.0",
     "@typescript-eslint/parser": "^2.28.0",
     "allure-commandline": "^2.13.0",
@@ -67,9 +69,11 @@
     "lint-staged": "^10.1.3",
     "normalize-newline": "^3.0.0",
     "prettier": "^2.0.4",
+    "rollup": "^2.13.1",
+    "rollup-plugin-typescript2": "^0.27.1",
     "semantic-release": "^17.0.4",
     "testcafe": "^1.8.4",
     "ts-jest": "^25.5.1",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.5"
   }
 }
