--- tests/utils/test-callsite.js
@@ -1,7 +1,0 @@
-const createCallsiteRecord = require('callsite-record');
-
-try {
-  throw new Error('Create error');
-} catch (err) {
-  module.exports = createCallsiteRecord({ forError: err });
-}
