+++ tests/utils/test-callsite.ts
@@ -1,0 +1,9 @@
+const createCallsiteRecord = require('callsite-record');
+
+export default function () {
+  try {
+    throw new Error('Create error');
+  } catch (err) {
+    return createCallsiteRecord({ forError: err });
+  }
+}
