+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
     "compilerOptions": {
-        "module": "commonjs",
-        "outDir": "dist",
         "downlevelIteration": true,
         "moduleResolution": "Node",
+        "allowSyntheticDefaultImports": true,
         "typeRoots": [
             "node_modules/@types",
             "src/types"
