+++ tsconfig.prod.json
@@ -1,7 +1,7 @@
 {
-    "extends": "./tsconfig.base.json",
+    "extends": "./tsconfig.json",
     "compilerOptions": {
-        "declaration": true,
+        "declaration": true
     },
     "include": [
         "src/**/*.ts",
